+++ src/commonMain/kotlin/arduino/ArduinoProtocol.kt
@@ -44,9 +44,10 @@ class ArduinoProtocol(override val connection: Connection) : Protocol {
             add("  if (-1 < procedureIndex && Serial.available() >= bytesNeeded) callable = true;")
             add("}")
             add("")
-            add("void (*procedures[])() = {                           // index, hex , arguments")
+            val longestSyntax = procedures.maxOf { it.syntax.length }
+            add("${"void (*procedures[])() = { ".padEnd(longestSyntax)}            // index, hex , arguments")
             addAll(procedures.map {
-                "  [] { ${it.syntax.padEnd(41)}; }, //  #${it.index.toString().padStart(3)}, 0x${it.procedure.hex}, ${it.argBytes} b"
+                "  [] { ${it.syntax.padEnd(longestSyntax)}; }, //  #${it.index.toString().padStart(3)}, 0x${it.procedure.hex}, ${it.argBytes} b"
             })
             add("};")
             add("")
@@ -63,40 +64,48 @@ class ArduinoProtocol(override val connection: Connection) : Protocol {
     private fun getMicros(function: String, argBytes: Int, vararg args: Any) =
         add(GetMicros(send4(function, *args), argBytes))
 
+    private fun setByte(function: String, vararg args: Any) = add(SetByte(function(function, *args)))
+
     //#region Protocol implementation
     override fun analogRead(pin: String) = add(GetUShort(send(2, "analogRead", pin)))
 
     override fun analogReference(reference: Reference) = call("analogReference", reference.name.uppercase())
 
-    override fun analogWrite(pin: String) = add(SetByte(function("analogWrite", pin, READ_BYTE)))
+    override fun analogWrite(pin: String) = setByte("analogWrite", pin.pin, READ_BYTE)
 
-    override fun digitalRead(pin: String) = add(GetBoolean(send(1, "digitalRead", pin)))
+    override fun dallasCelsius() = add(ByteToCelsius(
+        "dallas.requestTemperatures(); send2(dallas.getTempCByIndex($READ_BYTE) * 100.0)"))
 
-    override fun digitalWrite(pin: String, high: Boolean) = call("digitalWrite", pin, high.constant)
+    override fun digitalRead(pin: String) = add(GetBoolean(send(1, "digitalRead", pin.pin)))
+
+    override fun digitalWrite(pin: String, high: Boolean) = call("digitalWrite", pin.pin, high.pad)
 
     override fun interrupts(enable: Boolean) = call("${if (enable) "i" else "noI"}nterrupts")
 
     override fun millis() = add(GetMillis(send4("millis")))
 
-    override fun noTone(pin: String) = call("noTone", pin)
+    override fun noTone(pin: String) = call("noTone", pin.pin)
 
-    override fun pinMode(pin: String, mode: Mode) = call("pinMode", pin, mode.arduino)
+    override fun pinMode(pin: String, mode: Mode) = call("pinMode", pin.pin, mode.pad)
 
-    override fun pulseIn(pin: String, high: Boolean) = getMicros("pulseIn", pin, high.constant)
+    override fun pulseIn(pin: String, high: Boolean) = getMicros("pulseIn", pin.pin, high.pad)
 
-    override fun pulseInLong(pin: String, high: Boolean) = getMicros("pulseInLong", pin, high.constant)
+    override fun pulseInLong(pin: String, high: Boolean) = getMicros("pulseInLong", pin.pin, high.pad)
 
     override fun pulseInLongTimeout(pin: String, high: Boolean) =
-        getMicros("pulseInLong", 4, pin, high.constant, READ_LONG)
+        getMicros("pulseInLong", 4, pin.pin, high.pad, READ_LONG)
 
     override fun pulseInTimeout(pin: String, high: Boolean) =
-        getMicros("pulseIn", 4, pin, high.constant, READ_LONG)
+        getMicros("pulseIn", 4, pin.pin, high.pad, READ_LONG)
 
     override fun reconnect() = add(IntToString("speed = $READ_LONG; Serial.end(); setup()"))
 
-    override fun tone(pin: String) = add(SetShort(function("tone", pin, READ_INT)))
+    override fun tone(pin: String) = add(SetShort(function("tone", pin.pin, READ_INT)))
 
-    override fun toneDuration(pin: String) = add(SetShortAndInt(function("tone", pin, READ_INT, READ_LONG)))
+    override fun toneDuration(pin: String) = add(SetShortAndInt(function("tone", pin.pin, READ_INT, READ_LONG)))
+
+    override fun wire1() =
+        setByte("one = OneWire($READ_BYTE); dallas = DallasTemperature(&one); dallas.begin")
     //#endregion
 
     companion object {
@@ -119,8 +128,13 @@ class ArduinoProtocol(override val connection: Connection) : Protocol {
         """.trimIndent()
 
         private val template = """
+            #include <OneWire.h> 
+            #include <DallasTemperature.h>
+            
             int bytesNeeded = 0;        // How many bytes procedure needs to read the arguments.
             boolean callable = false;   // Can the procedure be called.
+            DallasTemperature dallas;   // Dallas digital temperature sensors.
+            OneWire one;                // 1-wire communication bus.
             int procedureIndex = -1;    // The procedure index in lambda array.
             long speed = 9600L;         // Serial speed in bauds.
             
@@ -166,6 +180,16 @@ class ArduinoProtocol(override val connection: Connection) : Protocol {
 
         private fun send4(generator: String, vararg args: Any) = send(4, generator, *args)
 
-        private val Boolean.constant get() = if (this) "HIGH" else "LOW"
+        private val Boolean.pad get() = if (this) "HIGH" else "LOW "
+
+        @Suppress("SpellCheckingInspection")
+        private val Mode.pad get() = when (this) {
+            Mode.Input          -> "INPUT       "
+            Mode.InputPullUp    -> "INPUT_PULLUP"
+            Mode.Output         -> "OUTPUT      "
+        }
+
+        private val String.pin get() = padStart(2)
+
     }
 }