+++ src/jvmTest/kotlin/arduino/Dallas.kt
@@ -1,0 +1,12 @@
+package fi.papinkivi.crap.arduino
+
+fun main() = with (Uno("COM3")) {
+    val device = attachDallasTemperature()
+    var previous = 0f
+    while (connected) {
+        previous = device.celsius.apply {
+            if (previous != this)
+                println("Changed to $this °C.")
+        }
+    }
+}