+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-10-11
+### Added
+- Dallas Temperature sensors support.
+- Dallas example.
+- 1-wire support.
+- Module API.
+
+### Changed
+- Protocol procedures.
+- Shield API.
+
 ## [1.0.0] - 2022-10-10
 ### Added
 - Local Maven publish.